+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yourloops data visualization for diabetes device data 
 Library for Tidepool's timeline-style diabetes data visualization(s) used in Blip and was forked from tidepool/tideline.
 
+## unreleased
+### Added
+- [PT-508] Display reservoir changes in main page for Diabeloop devices.
+
 ## 1.0.0 - 2019-08-05
 ### Added
 - Integrate Tidepool changes until v1.12.0 https://github.com/tidepool-org/tideline/tree/v1.12.0