+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- PT-545 Display infusion site changes for Diabeloop Devices
+
 ### Fixed
 - PT-532 One can create a patient even if the application is not allowing it